Optimize runStringHandlers
I noticed that we loop over every declaration in this method, seeing if
there is a stringHandler available for it. However, it is more likely
that there will be fewer stringHandlers than declarations, so we can
speed this up by inverting the loop and check here.

While I was at it, I decided to prevent this method from creating a new
OrderedElements object every time and instead to just mutate it. I
believe this will be safe and should buy us more performance. This means
we no longer need the .map method on this class.

This change reduces the time spent in runStringHandlers from 8.6% of
css() runtime to 0.26% and brings the runtime of my benchmark down from
2150ms to 1950ms.